@@ -5,13 +5,19 @@ require (
 	github.com/clmul/checksum v0.1.0
 	github.com/clmul/socks5 v0.0.0-20180327061726-1a1592f2b65e
 	github.com/clmul/water v0.0.2
-	github.com/google/btree v0.0.0-20180813153112-4030bb1f1f0c // indirect
-	github.com/google/go-cmp v0.2.0
-	github.com/google/netstack v0.0.0-00010101000000-000000000000
-	github.com/miekg/dns v1.1.35 // indirect
-	golang.org/x/sys v0.0.0-20191104094858-e8c54fb511f6
+	github.com/google/go-cmp v0.5.6
+	github.com/miekg/dns v1.1.49
+	golang.org/x/sys v0.0.0-20220610221304-9f5ed59c137d
+	gvisor.dev/gvisor v0.0.0-20220611072857-67ec4b4c7749
 )
 
-replace github.com/google/netstack => github.com/clmul/netstack v0.0.0-20190308035238-c320e3f68db0
+require (
+	github.com/google/btree v1.0.1 // indirect
+	golang.org/x/mod v0.6.0-dev.0.20220419223038-86c51ed26bb4 // indirect
+	golang.org/x/net v0.0.0-20220607020251-c690dde0001d // indirect
+	golang.org/x/time v0.0.0-20191024005414-555d28b269f0 // indirect
+	golang.org/x/tools v0.1.11 // indirect
+	golang.org/x/xerrors v0.0.0-20220609144429-65e65417b02f // indirect
+)
 
-go 1.13
+go 1.18
