@@ -10,33 +10,35 @@ import (
 	"github.com/clmul/cutevpn"
 )
 
-type udp4 struct {
-	listen string
+type udp struct {
+	cipher cutevpn.Cipher
 	peer   cutevpn.LinkAddr
 	conn   *net.UDPConn
 	ctx    context.Context
 	cancel context.CancelFunc
 }
 
-func newUDP(vpn cutevpn.VPN, ctx context.Context, linkURL *url.URL) (cutevpn.Link, error) {
-	ctx, cancel := context.WithCancel(ctx)
-	t := &udp4{
+func newUDP(vpn cutevpn.VPN, linkURL *url.URL, cipher cutevpn.Cipher) error {
+	ctx, cancel := context.WithCancel(vpn.Context())
+	t := &udp{
+		cipher: cipher,
 		ctx:    ctx,
 		cancel: cancel,
 	}
 
-	if linkURL.Query().Get("listen") == "1" {
-		t.listen = linkURL.Host
+	var listen string
+	if linkURL.Hostname() == "" {
+		listen = linkURL.Host
 	} else {
-		addr, err := net.ResolveUDPAddr("udp4", linkURL.Host)
+		addr, err := net.ResolveUDPAddr("udp", linkURL.Host)
 		if err != nil {
-			return nil, err
+			return err
 		}
 		t.peer = cutevpn.ConvertNetAddr(addr.IP, addr.Port)
 	}
-	c, err := net.ListenPacket("udp4", t.listen)
+	c, err := net.ListenPacket("udp", listen)
 	if err != nil {
-		return nil, err
+		return err
 	}
 	vpn.OnCancel(ctx, func() {
 		err := c.Close()
@@ -45,43 +47,50 @@ func newUDP(vpn cutevpn.VPN, ctx context.Context, linkURL *url.URL) (cutevpn.Lin
 		}
 	})
 	t.conn = c.(*net.UDPConn)
+	vpn.AddLink(t)
 
-	return t, nil
+	return nil
 }
 
-func (t *udp4) ToString(dst cutevpn.LinkAddr) string {
+func (t *udp) ToString(dst cutevpn.LinkAddr) string {
 	if dst == nil {
 		dst = "any"
 	}
-	return fmt.Sprintf("udp4 %v->%v", t.conn.LocalAddr(), dst)
+	return fmt.Sprintf("udp %v->%v", t.conn.LocalAddr(), dst)
 }
 
-func (t *udp4) Peer() cutevpn.LinkAddr {
+func (t *udp) Peer() cutevpn.LinkAddr {
 	return t.peer
 }
 
-func (t *udp4) Send(packet []byte, addr cutevpn.LinkAddr) error {
+func (t *udp) Send(packet []byte, addr cutevpn.LinkAddr) error {
 	ip, port := cutevpn.ConvertToNetAddr(addr.(cutevpn.AddrPort))
-	_, err := t.conn.WriteToUDP(packet, &net.UDPAddr{IP: ip, Port: port})
+	_, err := t.conn.WriteToUDP(t.cipher.Encrypt(packet), &net.UDPAddr{IP: ip, Port: port})
 	return err
 }
 
-func (t *udp4) Recv(packet []byte) (p []byte, addr cutevpn.LinkAddr, err error) {
+func (t *udp) Recv(packet []byte) (p []byte, addr cutevpn.LinkAddr, err error) {
 	n, udpAddr, err := t.conn.ReadFromUDP(packet)
 	if err != nil {
 		return nil, nil, err
 	}
-	return packet[:n], cutevpn.ConvertNetAddr(udpAddr.IP, udpAddr.Port), nil
+	packet = packet[:n]
+	packet, err = t.cipher.Decrypt(packet)
+	if err != nil {
+		log.Println(err)
+		return packet[:0], nil, nil
+	}
+	return packet, cutevpn.ConvertNetAddr(udpAddr.IP, udpAddr.Port), nil
 }
 
-func (t *udp4) Overhead() int {
-	return 20 + 8
+func (t *udp) Overhead() int {
+	return 20 + 8 + t.cipher.Overhead()
 }
 
-func (t *udp4) Cancel() {
+func (t *udp) Cancel() {
 	t.cancel()
 }
 
-func (t *udp4) Done() <-chan struct{} {
+func (t *udp) Done() <-chan struct{} {
 	return t.ctx.Done()
 }
