@@ -5,14 +5,13 @@ import (
 	"testing"
 )
 
-func init() {
-	bootTime = 127
-}
+var bootTime uint64 = 127
 
 func TestMarshalHello(t *testing.T) {
-	p0 := NewHello(IPv4{192, 168, 123, 234}, 1, 2, 3)
+	p0 := NewHello(1, 2, 3)
+	p0.Src = IPv4{192, 168, 123, 234}
 	p0.BootTime = bootTime
-	marshaled := p0.Marshal(make([]byte, 2048))
+	marshaled := p0.Marshal(make([]byte, 2048), p0.Src, p0.BootTime)
 	p1 := Unmarshal(marshaled)
 
 	if p0 != p1 {
@@ -21,9 +20,10 @@ func TestMarshalHello(t *testing.T) {
 }
 
 func TestMarshalLinkStateAck(t *testing.T) {
-	p0 := NewLinkStateACK(IPv4{1, 1, 1, 1}, IPv4{8, 8, 8, 8}, 1345245240)
+	p0 := NewLinkStateACK(IPv4{8, 8, 8, 8}, 1345245240)
 	p0.BootTime = bootTime
-	marshaled := p0.Marshal(make([]byte, 2048))
+	p0.Src = IPv4{1, 1, 1, 1}
+	marshaled := p0.Marshal(make([]byte, 2048), p0.Src, p0.BootTime)
 	p1 := Unmarshal(marshaled)
 
 	if p0 != p1 {
@@ -38,8 +38,9 @@ func TestMarshalLinkStateUpdate(t *testing.T) {
 		{3, 1, 1, 1}:   135246789,
 		{3, 1, 1, 255}: 1352467890,
 	})
+	p0.Src = IPv4{1, 1, 1, 1}
 	p0.BootTime = bootTime
-	marshaled := p0.Marshal(make([]byte, 2048))
+	marshaled := p0.Marshal(make([]byte, 2048), p0.Src, p0.BootTime)
 	p1 := Unmarshal(marshaled)
 
 	if fmt.Sprint(p0) != fmt.Sprint(p1) {
