@@ -10,7 +10,8 @@ func New(name string, vpn cutevpn.VPN, cidr string, mtu uint32) (cutevpn.Socket,
 	switch name {
 	case "tun":
 		return openTun(vpn, cidr, mtu)
-	//case "socks5":
+	case "socks5":
+		return openSocks5(vpn, cidr, mtu)
 	default:
 		return nil, fmt.Errorf("unknown socket %s", name)
 	}
